Updated the warning message for time range from unit tests
