+++ .github/workflows/pr_check.yml
@@ -31,6 +31,7 @@ jobs:
     - name: Compare with main branch
       run: |
         python -m unittest tests/test_getSQfield.py
+        python -m unittest tests/test_validate_inputs.py
         python tests/validate_xdifi2.py
         python tests/validate_difi8.py
-        python tests/validate_above_sq_current.py
+
